requirements.txt
transformers
sentence-transformers
datasets
torch
beautifulsoup4
transformers[torch]
scikit-learn
numpy
faiss-cpu  # or faiss-gpu
sentence-transformers
tqdm
sentencepiece